backend/requirements.txt
fastapi
uvicorn
speechrecognition
transformers
torch
numpy
pydantic
python-multipart
scipy
aiofiles
fastapi
uvicorn
speechrecognition
transformers
torch
pydantic
python-multipart
flask
flask_cors
openai-whisper
phonemizer
espeak-ng
pydub
